// src/app/shared/models/car.model.ts
export interface CarModel {
  remito: string;
  motor: string;
  fecha_patentamiento: string;
  ubicacion: number;
  estado: string;
  pvta_ori: string;
  color: string;
  factura_compra: string;
  publicadoMercadoLibre: string;
  polizaVto: string;
  polizaNro: string;
  posicion: number;
  auto_id: number;
  anio_produccion: number;
  marca: string;
  itemID: string;
  preventa: string;
  pedido: string;
  carroceria: string;
  vin: string;
  modelo_descripcion: string;
  certificado_anio: string;
  duenio: string;
  companiaDeSeguroID: string;
  garantiaKm: number;
  modelo: string;
  certificado: string;
  patente: string;
  tabla_automotor: string;
  color_id: number;
  garantiaFecha: string;
  autorizo_entrega: string;
  preventa_id: number;
  serie: string;
  autorizador_preventa: string;
  tapizado: string;
  id_fabrica: string;
  anio: string;
  referencia: string;
  valor_compra: number;
  kms: number;
  precioDeVenta: number;
  combustible: string;
  localidad: string;
  priceNumber: number;
  condicion: string;
}